requirements/base.txt
Django==4.2.17
djangorestframework==3.16.0
setuptools==82.0.0
gunicorn==25.0.0
psycopg2-binary==2.9.9
python-jose==3.3.0
PyJWT==2.10.1
cryptography==45.0.5
requests==2.32.4
httpx==0.28.1

# Procesamiento de documentos
python-docx==1.1.0
pandas==2.2.2
openpyxl==3.1.2
PyPDF2==3.0.1
drf-spectacular==0.28.0
# Vectorización y recuperación
#pgvector==0.2.3

# Embeddings
sentence-transformers==2.6.1
#transformers==4.38.2
#huggingface_hub==0.23.0
#torch==2.3.1+cpu
scikit-learn==1.4.2

#nomic
#sentence-transformers
# langchain
langchain-text-splitters
pgvector
#psycopg2-binary
python-dotenv
#textract
langdetect
python-magic
#torch
transformers
ollama
ijson

# Celery y Redis para tareas asíncronas
celery==5.3.4
redis==5.0.1
django-celery-results

# Geospatial libraries
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
# Generación de reportes
weasyprint
Markdown
python-pptx
geopandas
